refactor: reduce dom.js exports to core-only (_el, createActionButton, renderButtonBar, buildChevronRow)
Move _safeFit to terminal-factory.js, createModalOverlay to dom-dialogs.js,
createSelect to flow-modal-helpers.js, positionInViewport to context-menu.js.
